Rough skeleton of tinygo wasm build integration
This sets up a basic nix development environment, CI workflow, and integration
of a go tinygo wasm build with the upstream openai SDK. This is enough to
validate that the approach works with both node and browser, so next steps will
involve the actual integration with the proxy. I'm generally trying to keep the
changes from upstream as minimal as possible to make merging in upstream
changes easier.


Merges #1

LGTM given by: @jason-sindri