Integrate wasm build with openai client
The openai client now automatically points to the sindri backend and uses the
proxy through wasm to manage the attestation and encryption.

This switches the wasm build back to using the go compiler instead of tinygo
because tinygo can't handle protobufs and go-tdx-guest uses them heavily.
Something noteworthy is that the go wasm compiler typically disables fetch in
node because it assumes it's a testing environment for browser builds, so the
nix build here patches the go compiler to bypass this disabling.

Merges #2